chore: cleanup/refactor
