+++ src/Components/ServiceScene/Breakdowns/LayoutSwitcher.tsx
@@ -4,7 +4,7 @@ import { SelectableValue } from '@grafana/data';
 import { SceneComponentProps, SceneObject, SceneObjectBase, SceneObjectState } from '@grafana/scenes';
 import { Field, RadioButtonGroup } from '@grafana/ui';
 import { USER_EVENTS_ACTIONS, USER_EVENTS_PAGES, reportAppInteraction } from 'services/analytics';
-import { getSlug } from '../../../services/routing';
+import { getDrilldownSlug } from '../../../services/routing';
 
 export interface LayoutSwitcherState extends SceneObjectState {
   active: LayoutType;
@@ -28,7 +28,7 @@ export class LayoutSwitcher extends SceneObjectBase<LayoutSwitcherState> {
   public onLayoutChange = (active: LayoutType) => {
     reportAppInteraction(USER_EVENTS_PAGES.service_details, USER_EVENTS_ACTIONS.service_details.layout_type_changed, {
       layout: active,
-      view: getSlug(),
+      view: getDrilldownSlug(),
     });
     this.setState({ active });
   };
